+++ lib/components/led.ts
@@ -1,0 +1,8 @@
+import { commonComponentProps, lrPolarPins } from "lib/common/layout"
+import { z } from "zod"
+
+export const ledProps = commonComponentProps.extend({
+  color: z.string().optional(),
+})
+export const ledPins = lrPolarPins
+export type LedProps = z.input<typeof ledProps>
