+++ lib/components/silkscreen-text.ts
@@ -1,0 +1,13 @@
+import { length } from "circuit-json"
+import { pcbLayoutProps } from "lib/common/layout"
+import { z } from "zod"
+
+export const silkscreenTextProps = pcbLayoutProps.extend({
+  text: z.string(),
+  anchorAlignment: z
+    .enum(["center", "top_left", "top_right", "bottom_left", "bottom_right"])
+    .default("center"),
+  font: z.enum(["tscircuit2024"]).optional(),
+  fontSize: length.optional(),
+})
+export type SilkscreenTextProps = z.input<typeof silkscreenTextProps>
